backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.8.0
pydantic==2.12.5
python-dotenv==1.2.1
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.22
bcrypt==4.0.1
email-validator==2.3.0
starlette==0.37.2
cachetools==5.3.3

openpyxl==3.1.2
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.0.0
